+++ fpdf/table.py
@@ -120,6 +120,13 @@ class Table:
                     raise FPDFException(
                         f"Using font emphasis '{emphasis.style}' in table headings require the corresponding font style to be added using add_font()"
                     )
+        if self.rows:
+            cols_count = self.rows[0].cols_count
+            for i, row in enumerate(self.rows[1:], start=1):
+                if row.cols_count != cols_count:
+                    raise FPDFException(
+                        f"Inconsistent column count detected on row {i}"
+                    )
         # Defining table global horizontal position:
         prev_l_margin = self._fpdf.l_margin
         if table_align == Align.C:
@@ -203,27 +210,28 @@ class Table:
     def _render_table_row(self, i, row_layout_info, fill=False, **kwargs):
         row = self.rows[i]
         j = 0
-        while j < len(row.cells):
+        for cell in row.cells:
             self._render_table_cell(
                 i,
                 j,
+                cell,
                 row_height=row_layout_info.height,
                 fill=fill,
                 **kwargs,
             )
-            j += row.cells[j].colspan
+            j += cell.colspan
         self._fpdf.ln(row_layout_info.height)
 
     def _render_table_cell(
         self,
         i,
         j,
+        cell,
         row_height,
         fill=False,
         **kwargs,
     ):
         row = self.rows[i]
-        cell = row.cells[j]
         col_width = self._get_col_width(i, j, cell.colspan)
         if j and self._gutter_width:
             self._fpdf.x += self._gutter_width
@@ -296,6 +304,8 @@ class Table:
         for k in range(j, j + colspan):
             col_ratio = self._col_widths[k] / sum(self._col_widths)
             col_width += col_ratio * width
+            if k != j:
+                col_width += self._gutter_width
         return col_width
 
     def _get_row_layout_info(self, i):
@@ -308,12 +318,15 @@ class Table:
         any_page_break = False
         # pylint: disable=protected-access
         with self._fpdf._disable_writing():
-            for j in range(len(row.cells)):
+            j = 0
+            for cell in row.cells:
                 page_break, height = self._render_table_cell(
                     i,
                     j,
+                    cell,
                     row_height=self._line_height,
                 )
+                j += cell.colspan
                 any_page_break = any_page_break or page_break
                 heights_per_cell.append(height)
         row_height = (